### What does this PR do?
more parallelization on build steps
ex of current action https://github.com/salesforcecli/sfdx-cli/actions/runs/3490663206/jobs/5842412691


### What issues does this PR fix or reference?
@W-11726535@

### perf changes

smoke test on local mac

before. ../plugin-release-management/bin/dev cli:tarballs:smoke --cli sfdx  316.05s user 92.76s system 108% cpu 6:15.10 total
after. ../plugin-release-management/bin/dev cli:tarballs:smoke --cli sfdx  395.89s user 124.92s system 563% cpu 1:32.34 total

### QA setup

1. clone sfdx-cli, yarn install && yarn build
3. yarn pack:tarballs will get the tarballs prepared/packed (not required or impact by this change, but https://github.com/oclif/oclif/pull/1035 will save a lot of time)
4. either sfdx plugins:link this plugin, or use bin dev for the next 2 commands
5. `cli:tarballs:verify`
6. `cli:tarballs:smoke --cli sfdx`