@iowillhoit 2022-11-29
Looks awesome! Great improvement for the smoke command 🐇 

## `cli:tarballs:smoke`
BEFORE ../plugin-release-management/bin/dev cli:tarballs:smoke --cli sfdx  185.44s user 44.41s system 102% cpu 3:44.84 total
AFTER ../plugin-release-management/bin/dev cli:tarballs:smoke --cli sfdx  253.73s user 69.02s system 654% cpu 49.313 total

## `cli:tarballs:verify`
🟡 The output is getting a little garbled on this command. Notice that the status of 2 through 9 is not listed and some other output is out of order. The execution time of `verify` sync vs async is almost the same. Might be worth switching back to sync for the clean output. 
BEFORE
```
> time ../plugin-release-management/bin/dev cli:tarballs:verify --cli sfdx
[1/9] Ensure webdriverio does not exist... PASSED
[2/9] Ensure heroku-cli-util/.nyc_output does not exist... PASSED
Windows Path Length Test:
  - max windows path length: 259
  ---- Upper Limit ----
  - sfdx max username length: 64
  - sfdx max base path length: 136
  - sfdx max allowable path length: 123
  ---- Supported Limit ----
  - sfdx supported username length: 41
  - sfdx supported base path length: 113
  - sfdx supported allowable path length: 146
WARNING: Some paths could result in errors for Windows users with usernames that are 64 characters!
124 - node_modules/@salesforce/templates/lib/templates/analytics/waveTemplates/DefaultAnalyticsTemplate/app-to-template-rules.json
128 - node_modules/@salesforce/templates/lib/templates/analytics/waveTemplates/DefaultAnalyticsTemplate/dashboards/basicDashboard.json
124 - node_modules/@salesforce/templates/lib/templates/analytics/waveTemplates/DefaultAnalyticsTemplate/template-to-app-rules.json
127 - node_modules/@salesforce/templates/lib/templates/lightningcomponent/lwc/analyticsDashboard/__tests__/analyticsDashboard.test.js
143 - node_modules/@salesforce/templates/lib/templates/lightningcomponent/lwc/analyticsDashboardWithStep/__tests__/analyticsDashboardWithStep.test.js
130 - node_modules/@salesforce/templates/lib/templates/lightningcomponent/lwc/analyticsDashboardWithStep/analyticsDashboardWithStep.html
128 - node_modules/@salesforce/templates/lib/templates/lightningcomponent/lwc/analyticsDashboardWithStep/analyticsDashboardWithStep.js
137 - node_modules/@salesforce/templates/lib/templates/lightningcomponent/lwc/analyticsDashboardWithStep/analyticsDashboardWithStep.js-meta.xml
[3/9] Ensure windows path lengths... PASSED
[4/9] Ensure apex-node exists... PASSED
[5/9] Ensure plugin generator test template exists... PASSED
[6/9] Ensure templates commands exist... PASSED
[7/9] Ensure no tests or maps in dist... PASSED
[8/9] Ensure no unexpected files... PASSED
[9/9] Ensure sf is included
...
[HELP TEXT OMITTED]
[9/9] Ensure sf is included
... PASSED
../plugin-release-management/bin/dev cli:tarballs:verify --cli sfdx  3.12s user 1.97s system 143% cpu 3.531 total
```

AFTER
```
> time ../plugin-release-management/bin/dev cli:tarballs:verify --cli sfdx
Windows Path Length Test:
  - max windows path length: 259
  ---- Upper Limit ----
  - sfdx max username length: 64
  - sfdx max base path length: 136
  - sfdx max allowable path length: 123
  ---- Supported Limit ----
  - sfdx supported username length: 41
  - sfdx supported base path length: 113
  - sfdx supported allowable path length: 146
[1/9] Ensure sf is included
... PASSED
[HELP TEXT OMITTED]

WARNING: Some paths could result in errors for Windows users with usernames that are 64 characters!
124 - node_modules/@salesforce/templates/lib/templates/analytics/waveTemplates/DefaultAnalyticsTemplate/app-to-template-rules.json
128 - node_modules/@salesforce/templates/lib/templates/analytics/waveTemplates/DefaultAnalyticsTemplate/dashboards/basicDashboard.json
124 - node_modules/@salesforce/templates/lib/templates/analytics/waveTemplates/DefaultAnalyticsTemplate/template-to-app-rules.json
127 - node_modules/@salesforce/templates/lib/templates/lightningcomponent/lwc/analyticsDashboard/__tests__/analyticsDashboard.test.js
143 - node_modules/@salesforce/templates/lib/templates/lightningcomponent/lwc/analyticsDashboardWithStep/__tests__/analyticsDashboardWithStep.test.js
130 - node_modules/@salesforce/templates/lib/templates/lightningcomponent/lwc/analyticsDashboardWithStep/analyticsDashboardWithStep.html
128 - node_modules/@salesforce/templates/lib/templates/lightningcomponent/lwc/analyticsDashboardWithStep/analyticsDashboardWithStep.js
137 - node_modules/@salesforce/templates/lib/templates/lightningcomponent/lwc/analyticsDashboardWithStep/analyticsDashboardWithStep.js-meta.xml
../plugin-release-management/bin/dev cli:tarballs:verify --cli sfdx  3.02s user 1.93s system 144% cpu 3.437 total
```